doc: fix list formatting in gitformat-index
The document relies on indentation of the list items to denote a
nested list, but this isn't how asciidoc represents them. In asciidoc, to do
a nested list, one uses an increasing number of `*` characters, so the
top level uses `*` (`-` is acceptable), the first nested list uses `**`
(`--` isn't acceptable), the second nested list uses `***`, etc.

Switch the list character from `-` to `*`, and use `**` where a nested
list is requested.

The only changes in the rendering in `man` on my Linux machine involved
proper nesting of the sublists.