doc: make description of index entries a list
When rendered in `man` or HTML, the description of the items composing
an index entry are displayed without bullets. The txt file relies on
indentation to delineate a sublist, such as the ones describing the mode
or flag bits. This doesn't work with asciidoc, which interprets each
block individually, and treats one-or-more spaces as a request for
preformatting of that paragraph, but doesn't include those spaces
themselves. Example:

```
Normal Block

  Preformatted block (indented, line breaks respected)

              Also a preformatted block (indented the same amount)
```

Add list markers to the list-like data that doesn't currently have them.
Use a line-terminal `+` to force a line break. Use a list continuation
(`+` on its own line immediately following a list item) to properly
attach multiple paragraphs to a list item.

This causes some formatting changes when rendered in `man` and in the
browser. Aside from the additional bullets denoting the list items, this
changes the formatting a bit for multi-line list items. Previously, they
were rendered like this in `man`:

```
32-bit ctime seconds, the last time a file's metadata changed
  this is stat(2) data

32-bit ctime nanosecond fractions
  this is stat(2) data
```

They are now rendered like this in `man`, with the extra blank line
between the first line of the list item and the following lines:

```
*   32-bit ctime seconds, the last time a file's metadata changed

    this is stat(2) data

*   32-bit ctime nanosecond fractions

    this is stat(2) data
```

Additionally, line wrapping is no longer following the pre-formatted
line wrapping in the file, it's up to the renderer to line wrap. I
believe this is an improvement, as the line wrapping was previously not
semantic except in the cases that I added a forced line break.

Finally, the paragraph at the end of the index entry description,
starting with "Interpretation of index entries", becomes indented by an
additional level. I'm not sure why this is happening, but it's
resolvable by not relying on preformatting (that is, by removing the
indentation), so there's no attempt to resolve this now. This paragraph
joins three others later on in the document to have an extraneous
indent, all will be resolved in the next commit.