doc: stop preformatting gitformat-index
Paragraphs with the first line indented by at least one space are
considered "Literal Paragraphs" and the formatting changes to be
essentially preformatted text. This means that line breaks are
respected and rendered as-is.

Additionally, these literal paragraphs are rendered with
an increased indentation level in `man`, and when rendered in HTML are
surrounded by `<pre><code>` tags. This causes them to be displayed in a
monospace font, potentially in a different color. The change from
monospace font in one color to the variable-width font and other color
used by the bulleted lists is distracting, even more so with the
formatting in use at https://git-scm.com/docs/gitformat-index. It also
makes the `linkgit:gitformat-pack[5]` link not work as intended.

Remove the extraneous indentation at the beginning of each line to stop
triggering this "Literal Paragraph" formatting. Add blank lines as
necessary to preserve semantic whitespace. Add backtics to request
monospace/code formatting (and suppress some other formatting) where
appropriate.